Add Barcode scanner for product search Fixes #20
