+++ frontend/.eslintrc.json
@@ -6,7 +6,8 @@
     "extends": [
         "eslint:recommended",
         "plugin:@typescript-eslint/recommended",
-        "plugin:react/recommended"
+        "plugin:react/recommended",
+        "plugin:react/jsx-runtime"
     ],
     "parser": "@typescript-eslint/parser",
     "parserOptions": {
@@ -18,5 +19,6 @@
         "react"
     ],
     "rules": {
+        "react/prop-types": 0
     }
 }
