+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "vite-template-redux",
+  "name": "expiration-tracker-frontend",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.8.1",
+    "@yudiel/react-qr-scanner": "^1.1.10",
     "antd": "^5.8.6",
     "dayjs": "^1.11.9",
     "jwt-decode": "^3.1.2",
@@ -59,19 +60,5 @@
     "typescript": "*",
     "vite": "^4.0.0",
     "vitest": "^0.30.1"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ],
-    "plugins": [
-      "prettier"
-    ],
-    "rules": {
-      "prettier/prettier": "error",
-      "react/jsx-no-target-blank": "off"
-    }
-  },
-  "prettier": "prettier-config-nick"
+  }
 }
