+++ frontend/src/components/types.d.ts
@@ -1,0 +1,4 @@
+interface ParentModalProps {
+  handleClose?: () => void;
+  msg?: MessageInstance;
+}
