+++ frontend/src/pages/ItemPage.tsx
@@ -1,14 +1,13 @@
 import { Layout } from "antd"
-import React from "react"
-import ItemPage from "../components/ItemTable"
+import ItemTable from "../components/ItemTable"
 import Header from "../components/common/Header"
 import Footer from "../components/common/Footer"
 
-export default function HomePage() {
+export default function ItemPage() {
   return (
     <Layout style={{ minHeight: "100vh" }}>
       <Header />
-      <ItemPage />
+      <ItemTable />
       <Footer />
     </Layout>
   )
