+++ frontend/src/pages/LoginPage.tsx
@@ -1,4 +1,3 @@
-import React from "react"
 import LoginForm from "../components/LoginForm"
 
 export default function LoginPage() {
